+++ interspace/include/interspace/world/EntityData.hpp
@@ -9,7 +9,7 @@ namespace Interspace
 {
     struct EntityData
     {
-        uint16_t id = 0;
+        uint32_t id = 0;
         std::string name;
         Engine::Vec2<float> position{};
         uint16_t sight = 32;
